requirements.txt
Babel==2.7.0
Django==2.2.5
django-countries==5.5
django-phonenumber-field==3.0.1
django-phonenumbers==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
phonenumbers==8.10.19
psycopg2==2.8.3
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0